chore(fv): Export BinderX type
We import the BinderX type into the Vir ast file because you need it to
generate ExprX::Ctor. If it's not imported you would need to add the Air
package to your project to be able to generate tuple related Vir
expressions. This will result in your project having all dependencies
related to the Air package which is burdensome.

With this type export this problem is being avoided.